apply corrections suggested by Niko on toString methods for M3 elements
