+++ core/src/main/java/io/lionweb/lioncore/java/language/Containment.java
@@ -99,11 +99,6 @@ public class Containment extends Link<Containment> {
   }
 
   @Override
-  public String toString() {
-    return super.toString() + "{" + "name=" + getName() + ", " + "type=" + getType() + '}';
-  }
-
-  @Override
   public Concept getConcept() {
     return LionCore.getContainment();
   }
