+++ core/src/main/java/io/lionweb/lioncore/java/language/Property.java
@@ -73,7 +73,7 @@ public class Property extends Feature<Property> {
 
   @Override
   public String toString() {
-    return super.toString() + "{" + "name=" + getName() + ", " + "type=" + getType() + '}';
+    return super.toString() + "{" + "qualifiedName=" + DebugUtils.qualifiedName(this) + ", " + "type=" + getType() + '}';
   }
 
   @Override
