+++ core/src/main/java/io/lionweb/lioncore/java/language/Reference.java
@@ -117,11 +117,6 @@ public class Reference extends Link<Reference> {
   }
 
   @Override
-  public String toString() {
-    return super.toString() + "{" + "name=" + getName() + ", " + "type=" + getType() + '}';
-  }
-
-  @Override
   public Concept getConcept() {
     return LionCore.getReference();
   }
