+++ core/src/main/java/io/lionweb/lioncore/java/model/impl/M3Node.java
@@ -312,4 +312,5 @@ public abstract class M3Node<T extends M3Node> implements Node {
       referenceValues.put(linkName, new ArrayList(Arrays.asList(value)));
     }
   }
+
 }
